+++ src/xrGame/game_sv_mp.h
@@ -33,7 +33,7 @@ class game_sv_mp :public game_sv_GameState
 	typedef game_sv_GameState inherited;
 
 protected:
-	//список трупов для удаления
+	//СЃРїРёСЃРѕРє С‚СЂСѓРїРѕРІ РґР»СЏ СѓРґР°Р»РµРЅРёСЏ
 	using CORPSE_LIST = xr_deque<u16>;
 	using CORPSE_LIST_it = CORPSE_LIST::iterator;
 
@@ -123,8 +123,8 @@ public:
 
 	virtual		void				net_Export_State		(NET_Packet& P, ClientID id_to);
 
-	virtual		void				OnRoundStart			();												// старт раунда
-	virtual		void				OnRoundEnd				();	//round_end_reason							// конец раунда
+	virtual		void				OnRoundStart			();												// СЃС‚Р°СЂС‚ СЂР°СѓРЅРґР°
+	virtual		void				OnRoundEnd				();	//round_end_reason							// РєРѕРЅРµС† СЂР°СѓРЅРґР°
 	virtual		bool				OnNextMap				();
 	virtual		void				OnPrevMap				();
 	
@@ -158,7 +158,7 @@ public:
 	virtual		void				ConsoleCommands_Create	();
 	virtual		void				ConsoleCommands_Clear	();
 
-	virtual		u32					GetTeamCount			()	{return TeamList.size();};	
+	virtual		u32					GetTeamCount			()	{return (u32)TeamList.size();};
 				TeamStruct*			GetTeamData				(u32 Team);
 
 	virtual		u8					GetSpectatorModes		() {return m_u8SpectatorModes;};
